LQPI/requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.15.0
bitsandbytes>=0.41.0  # For quantization support
safetensors>=0.4.0

# Data processing and analysis
scikit-learn>=1.3.2
scipy>=1.11.3
pandas>=2.1.1

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
networkx>=3.2.1  # For graph analysis and visualization
plotly>=5.18.0   # For interactive visualizations

# Machine learning extensions
umap-learn>=0.5.5  # For dimensionality reduction
hdbscan>=0.8.33   # For clustering

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.1
psutil>=5.9.6     # For memory monitoring
typer>=0.9.0      # For command-line interfaces
tenacity>=8.2.3   # For retry logic
joblib>=1.3.2     # For parallel processing

# Development tools
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Optional dependencies for advanced visualizations
pillow>=10.1.0
imageio>=2.31.5   # For creating animations

peft